requirements.txt
matplotlib==3.4.2
torchaudio==0.9.0
numpy==1.21.1
torchvision==0.10.0+cpu
pandas==1.3.1
torch==1.9.0+cpu
Pillow==8.3.1
scikit_learn==0.24.2
